+++ garbage_maze.h
@@ -3,9 +3,9 @@
 
 #include <vector>
 
-void ensureConnectivity(std::vector<std::vector<int>>& grid);
-void placeEmptyBlock(std::vector<std::vector<int>>& grid, int blockWidth, int blockHeight);
-std::vector<std::vector<int>> generateMaze(int rows, int cols, int density);
+void ensureConnectivity(std::vector<std::vector<short int>>& grid);
+void placeEmptyBlock(std::vector<std::vector<short int>>& grid, int blockWidth, int blockHeight);
+std::vector<std::vector<short int>> generateMaze(int rows, int cols, int density);
 
 #endif
